backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
black
